third_party/WebKit/LayoutTests/accessibility/clickable-expected.txt

This tests the clickable property of various elements, and tests that a role of an anchor element changes to link when it has a click listener.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS isAXElementClickable('link') is true
PASS isAXElementClickable('button') is true
PASS isAXElementClickable('labeled-button') is true
PASS isAXElementClickable('button-with-title') is true
PASS isAXElementClickable('switch') is true
PASS isAXElementClickable('disclosure-triangle') is true
PASS isAXElementClickable('text') is true
PASS isAXElementClickable('search-box') is true
PASS isAXElementClickable('checkbox') is true
PASS isAXElementClickable('number') is true
PASS isAXElementClickable('radio') is true
PASS isAXElementClickable('slider') is false
PASS isAXElementClickable('submit') is true
PASS isAXElementClickable('combobox') is true
PASS isAXElementClickable('listbox-single') is true
PASS isAXElementClickable('listbox-multiple') is true
PASS isAXElementClickable('textarea') is true
PASS isAXElementClickable('content-editable') is true
PASS isAXElementClickable('focusable') is false
PASS isAXElementClickable('heading') is false
PASS isAXElementClickable('aria-button') is true
PASS isAXElementClickable('aria-link') is true
PASS isAXElementClickable('div') is false
PASS isAXElementClickable('onclick') is true
PASS isAXElementClickable('keydown-listener') is false
PASS isAXElementClickable('click-listener') is true
PASS isAXElementClickable('mousedown-listener') is true
PASS isAXElementClickable('click-listener-on-ancestor') is false
PASS axRole('ancestor-with-click-listener') is 'AXRole: AXGenericContainer'
PASS axRole('empty-anchor') is 'AXRole: AXAnchor'
PASS axRole('href-anchor') is 'AXRole: AXLink'
PASS axRole('onclick-anchor') is 'AXRole: AXLink'
PASS axRole('click-listener-anchor') is 'AXRole: AXLink'
PASS successfullyParsed is true

TEST COMPLETE
